Fish man was here
-Added a threading method based on Unity's Job system (thanks for the suggestion Bradson!)
This system seems to perform better on my pc than the traditional Task-based system I wrote earlier, I'll likely just ditch the experimental threading option and make this an optional toggle
-Optimized a few things in the tick loop, namely the neighbor calculations.
The performance analyzer showed that a signifcant amount of time was spent on the IntVec3 calculations, which weren't actually necessary. I inlined the InBounds() check and made it directly compute the index instead of adding two vectors and then converting them to their index form. Thousands of times.